Revert "Reduce offset"
This reverts commit a32d677831589eef6494f57c1fab711da707d53b.

Archipelago.MultiClient.Net uses longs to clear locations.

+++ RhythmDoctor.Archipelago/World/Structures/Item.cs
@@ -6,7 +6,7 @@ public struct Item
   public string Name;
 
   [YamlMember(Alias = "id")]
-  public uint ID;
+  public ulong ID;
 
   [YamlMember(Alias = "classification")]
   public ItemClassification Classification;
